rename FirstClassCallableRector to ArrayToFirstClassCallableRector to make feature explicit
