+++ rules-tests/Php81/Rector/Array_/ArrayToFirstClassCallableRector/Fixture/skip_in_attribute_class.php.inc
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\Fixture;
+
+#[SomeAttribute(['array-argument'])]
+class SkipInAttributeClass
+{
+}
