+++ rules-tests/Php81/Rector/Array_/ArrayToFirstClassCallableRector/Fixture/skip_invoke_from_non_class_code.php.inc
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\Fixture;
+
+use Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\Source\SomeExternalObject;
+
+Route::get('verify-email', [SomeExternalObject::class, '__invoke'])
+    ->name('verification.notice');
