+++ rules-tests/Php81/Rector/Array_/ArrayToFirstClassCallableRector/Fixture/skip_non_static_private_method_with_other_object_const_fetch.php.inc
@@ -1,8 +1,8 @@
 <?php
 
-namespace Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Fixture;
+namespace Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\Fixture;
 
-use Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Source\SomeExternalObject;
+use Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\Source\SomeExternalObject;
 
 final class SkipNonStaticPrivateMethodWithOtherObjectConstFetch
 {
