+++ rules-tests/Php81/Rector/Array_/ArrayToFirstClassCallableRector/Fixture/skip_on_class_const.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Fixture;
+namespace Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\Fixture;
 
 final class SkipOnClassConst
 {
