+++ rules-tests/Php81/Rector/Array_/ArrayToFirstClassCallableRector/Fixture/skip_private_method_by_instantiation.php.inc
@@ -1,8 +1,8 @@
 <?php
 
-namespace Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Fixture;
+namespace Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\Fixture;
 
-use Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Source\WithPrivateMethod;
+use Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\Source\WithPrivateMethod;
 
 final class SkipPrivateMethodByInstantiation
 {
