+++ rules-tests/Php81/Rector/Array_/ArrayToFirstClassCallableRector/Source/WithPrivateMethod.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Source;
+namespace Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\Source;
 
 class WithPrivateMethod {
   /** @var callable */
