+++ rules-tests/Php81/Rector/Array_/ArrayToFirstClassCallableRector/config/configured_rule.php
@@ -3,11 +3,11 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Php81\Rector\Array_\FirstClassCallableRector;
+use Rector\Php81\Rector\Array_\ArrayToFirstClassCallableRector;
 use Rector\ValueObject\PhpVersion;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(FirstClassCallableRector::class);
+    $rectorConfig->rule(ArrayToFirstClassCallableRector::class);
 
     $rectorConfig->phpVersion(PhpVersion::PHP_81);
 };
