--- rules-tests/Php81/Rector/Array_/FirstClassCallableRector/Fixture/skip_dynamic_variable.php.inc
@@ -1,6 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Fixture;
-
-$variableName = 'x';
-${$variableName} = call_user_func_array([], []);
