--- rules-tests/Php81/Rector/Array_/FirstClassCallableRector/Fixture/skip_in_attribute_class.php.inc
@@ -1,8 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Fixture;
-
-#[SomeAttribute(['array-argument'])]
-class SkipInAttributeClass
-{
-}
