--- rules-tests/Php81/Rector/Array_/FirstClassCallableRector/Fixture/skip_invoke_from_non_class_code.php.inc
@@ -1,8 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Fixture;
-
-use Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\Source\SomeExternalObject;
-
-Route::get('verify-email', [SomeExternalObject::class, '__invoke'])
-    ->name('verification.notice');
