+++ rules/Php81/Rector/Array_/ArrayToFirstClassCallableRector.php
@@ -29,9 +29,9 @@ use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
 /**
  * @see RFC https://wiki.php.net/rfc/first_class_callable_syntax
- * @see \Rector\Tests\Php81\Rector\Array_\FirstClassCallableRector\FirstClassCallableRectorTest
+ * @see \Rector\Tests\Php81\Rector\Array_\ArrayToFirstClassCallableRector\ArrayToFirstClassCallableRectorTest
  */
-final class FirstClassCallableRector extends AbstractRector implements MinPhpVersionInterface
+final class ArrayToFirstClassCallableRector extends AbstractRector implements MinPhpVersionInterface
 {
     public function __construct(
         private readonly ArrayCallableMethodMatcher $arrayCallableMethodMatcher,
