+++ tests/Issues/AnnotationToAttributeFirstClassCallable/config/configured_rule.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 use Rector\Config\RectorConfig;
 use Rector\Php80\Rector\Class_\AnnotationToAttributeRector;
 use Rector\Php80\ValueObject\AnnotationToAttribute;
-use Rector\Php81\Rector\Array_\FirstClassCallableRector;
+use Rector\Php81\Rector\Array_\ArrayToFirstClassCallableRector;
 
 return RectorConfig::configure()
     ->withConfiguredRule(AnnotationToAttributeRector::class, [
         new AnnotationToAttribute('Symfony\Component\Serializer\Annotation\Context'),
     ])
-    ->withRules([FirstClassCallableRector::class]);
+    ->withRules([ArrayToFirstClassCallableRector::class]);
