+++ tests/Issues/AttributeAndArgValueRefresh/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Php81\Rector\Array_\FirstClassCallableRector;
+use Rector\Php81\Rector\Array_\ArrayToFirstClassCallableRector;
 use Rector\Tests\Issues\AttributeAndArgValueRefresh\Source\SetArrayArgValueRector;
 use Rector\Tests\Issues\AttributeAndArgValueRefresh\Source\SetArrayAttributeValueRector;
 
@@ -11,5 +11,5 @@ return RectorConfig::configure()
     ->withRules([
         SetArrayArgValueRector::class,
         SetArrayAttributeValueRector::class,
-        FirstClassCallableRector::class,
+        ArrayToFirstClassCallableRector::class,
     ]);
